Very large refactoring (see commit comments)
• All input parameters are now considered nonnull unless stated
otherwise. If a nil value is supplied to a value assumed to be nonnull,
then an exception is thrown.
• Arrays are now typecast so that the values are known.
• Removed OTRKitPrepareForApplicationTerminationNotification in favor
of listening for NSApplicationWillTerminateNotification
• Any API that is considered private is now prefixed by an underscore
to make this intent as clear as possible.
• Returned internal dispatch queue for OTR operations.
• Returned callback dispatch queue for delegate.
• Reorganized and renamed many methods and variables for greater
consistency and readability.
• Changed to dot syntax for properties
• Migrated dialogs to a better auto layout, layout which allows for any
length question and answers for authentication.
• [OTRKitAuthenticationDialogIncoming
showFingerprintConfirmation:username:accountName:protocol:] no longer
creates a new instance of OTRKitAuthenticationDialog until the user
actually answers Yes or No. This prevents the sheet blocking incoming
requests before the user has answered.
• Fixed incorrect condition in [OTRKit
isGeneratingKeyForAccountName:protocol] —
otrl_privkey_generate_cancelled() was not called when return value for
otrl_privkey_generate_start() is zero (0). This is a bug exclusive to
this fork. OTRKit has correct condition.